requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
joblib==1.3.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
schedule==1.2.1
gunicorn==21.2.0

